feat: add selected days filter to calendar
